tests/wpt/web-platform-tests/offscreen-canvas/shadows/2d.shadow.outside.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.shadow.outside</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.shadow.outside</h1>
<p class="desc">Shadows of shapes outside the visible area can be offset onto the visible area</p>


<script>
var t = async_test("Shadows of shapes outside the visible area can be offset onto the visible area");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.shadowColor = '#0f0';
ctx.shadowOffsetX = 100;
ctx.fillRect(-100, 0, 25, 50);
ctx.shadowOffsetX = -100;
ctx.fillRect(175, 0, 25, 50);
ctx.shadowOffsetX = 0;
ctx.shadowOffsetY = 100;
ctx.fillRect(25, -100, 50, 25);
ctx.shadowOffsetY = -100;
ctx.fillRect(25, 125, 50, 25);
_assertPixel(offscreenCanvas, 12,25, 0,255,0,255, "12,25", "0,255,0,255");
_assertPixel(offscreenCanvas, 87,25, 0,255,0,255, "87,25", "0,255,0,255");
_assertPixel(offscreenCanvas, 50,12, 0,255,0,255, "50,12", "0,255,0,255");
_assertPixel(offscreenCanvas, 50,37, 0,255,0,255, "50,37", "0,255,0,255");

t.done();

});
</script>
